Adding in UI for project

+++ Assets/Scripts/Components/Collectable.cs
@@ -3,5 +3,5 @@
 [GenerateAuthoringComponent]
 public struct Collectable : IComponentData
 {
-    public float Points;
+    public int Points;
 }
